Merge master into fork

+++ src/main/java/org/primefaces/event/SlideEndEvent.java
@@ -1,5 +1,5 @@
-/*
- * Copyright 2009-2014 PrimeTek.
+/**
+ * Copyright 2009-2018 PrimeTek.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,14 +20,14 @@ import javax.faces.component.behavior.Behavior;
 
 public class SlideEndEvent extends AbstractAjaxBehaviorEvent {
 
-    private int value;
+    private double value;
 
-	public SlideEndEvent(UIComponent component, Behavior behavior, int value) {
-		super(component , behavior);
+    public SlideEndEvent(UIComponent component, Behavior behavior, double value) {
+        super(component, behavior);
         this.value = value;
-	}
+    }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
